Refactor GemseoGenerator
